Struct serdegen::Options

pub(crate) struct Options {
    input: Option<PathBuf>,
    language: Language,
    target_source_dir: Option<PathBuf>,
    with_runtimes: Vec<Runtime>,
    module_name: Option<String>,
    serde_package_name: Option<String>,
    use_c_style_enums: bool,
}

Fields
input: Option<PathBuf>

Path to the YAML-encoded Serde formats.

language: Language

Language for code generation.

target_source_dir: Option<PathBuf>

Directory where to write generated modules (otherwise print code on stdout).

with_runtimes: Vec<Runtime>

Optional runtimes to install in the target_source_dir (if applicable). Also triggers the generation of specialized methods for each runtime.

module_name: Option<String>

Module name for the Serde formats installed in the target_source_dir. Rust crates may contain a version number separated with a colon, e.g. “test:1.2.0”. (By default, the installer will use version “0.1.0”.)

serde_package_name: Option<String>

Optional package name (Python) or module path (Go) where to find Serde runtime dependencies.

use_c_style_enums: bool

Translate enums without variant data (c-style enums) into their equivalent in the target language, if the target language and the generator code support them.